Refactor: Add EdfWriter class for high-level EDF/BDF file writing and update EdfIO include paths
