+++ include/edfio/EdfFile.hpp
@@ -17,6 +17,7 @@
 #include <memory>
 #include <stdexcept>
 #include <string>
+#include <utility>
 #include <vector>
 
 namespace edfio {
@@ -112,8 +113,8 @@ public:
     validateSignalIndex(signalIndex);
 
     auto const &sig = m_header.m_signals[signalIndex];
-    auto store = detail::CreateSignalSampleStore(*m_stream, m_header.m_general,
-                                                 sig);
+    auto store =
+        detail::CreateSignalSampleStore(*m_stream, m_header.m_general, sig);
 
     ProcessorSampleRecord<SampleType::Physical> proc(sig.m_detail.m_offset,
                                                      sig.m_detail.m_scaling);
@@ -138,8 +139,8 @@ public:
     validateSignalIndex(signalIndex);
 
     auto const &sig = m_header.m_signals[signalIndex];
-    auto store = detail::CreateSignalSampleStore(*m_stream, m_header.m_general,
-                                                 sig);
+    auto store =
+        detail::CreateSignalSampleStore(*m_stream, m_header.m_general, sig);
 
     // offset=0, scaling=1 yields the raw digital value.
     ProcessorSampleRecord<SampleType::Digital> proc(0.0, 1.0);
@@ -166,8 +167,8 @@ public:
       if (!sig.m_detail.m_isAnnotation)
         continue;
 
-      auto sigStore = detail::CreateSignalRecordStore(
-          *m_stream, m_header.m_general, sig);
+      auto sigStore =
+          detail::CreateSignalRecordStore(*m_stream, m_header.m_general, sig);
 
       int64_t drIdx = 0;
       for (auto it = sigStore.begin(); it != sigStore.end(); ++it, ++drIdx) {
@@ -209,7 +210,7 @@ public:
   [[nodiscard]] SignalSampleStore signalSampleStore(size_t signalIndex) const {
     validateSignalIndex(signalIndex);
     return detail::CreateSignalSampleStore(*m_stream, m_header.m_general,
-                                            m_header.m_signals[signalIndex]);
+                                           m_header.m_signals[signalIndex]);
   }
 
 private:
@@ -218,99 +219,14 @@ private:
 
   void validateSignalIndex(size_t signalIndex) const {
     if (signalIndex >= m_header.m_signals.size()) {
-      throw std::out_of_range(
-          "Signal index " + std::to_string(signalIndex) +
-          " out of range [0, " +
-          std::to_string(m_header.m_signals.size()) + ")");
+      throw std::out_of_range("Signal index " + std::to_string(signalIndex) +
+                              " out of range [0, " +
+                              std::to_string(m_header.m_signals.size()) + ")");
     }
   }
 
   std::unique_ptr<std::ifstream> m_stream;
   HeaderExam m_header;
-};
-
-/// High-level facade for writing EDF/BDF files.
-///
-/// Owns the underlying output stream.  Write data records sequentially after
-/// the header has been written.
-///
-/// Moveable but not copyable (owns an std::ofstream).
-class EdfWriter {
-public:
-  EdfWriter(const EdfWriter &) = delete;
-  EdfWriter &operator=(const EdfWriter &) = delete;
-  EdfWriter(EdfWriter &&) noexcept = default;
-  EdfWriter &operator=(EdfWriter &&) noexcept = default;
-
-  /// Create a new EDF/BDF file and write its header.
-  ///
-  /// @param path    Filesystem path for the new file.
-  /// @param header  Complete header exam to write.
-  /// @throws std::runtime_error if the file cannot be created.
-  [[nodiscard]] static EdfWriter create(const std::filesystem::path &path,
-                                        const HeaderExam &header) {
-    auto stream = std::make_unique<std::ofstream>(path, std::ios::binary);
-    if (!stream->is_open()) {
-      throw std::runtime_error(std::string("Cannot create file: ") +
-                               path.string());
-    }
-    WriteHeaderExam(*stream, header);
-    return EdfWriter(std::move(stream), header.m_general);
-  }
-
-  /// Write a single data record to the file.
-  ///
-  /// The data record count in the header is automatically updated when
-  /// close() is called (or the writer is destroyed).
-  void writeDataRecord(const Record<char> &record) {
-    if (m_stream && m_stream->is_open()) {
-      *m_stream << record;
-      ++m_recordsWritten;
-    }
-  }
-
-  /// Number of data records written so far.
-  [[nodiscard]] int64_t recordsWritten() const noexcept {
-    return m_recordsWritten;
-  }
-
-  /// Flush, patch the data-record count in the header, and close.
-  void close() {
-    if (m_stream && m_stream->is_open()) {
-      // EDF/BDF header layout: the "number of data records" field is 8 bytes
-      // starting at offset 236 (after version[8] + patient[80] +
-      // recording[80] + startDate[8] + startTime[8] + headerSize[8] +
-      // reserved[44] = 236).
-      static constexpr std::streamoff kDataRecordCountOffset = 236;
-      static constexpr size_t kFieldSize = 8;
-
-      m_stream->seekp(kDataRecordCountOffset, std::ios::beg);
-      auto countStr = std::to_string(m_recordsWritten);
-      countStr.resize(kFieldSize, ' ');
-      m_stream->write(countStr.data(), kFieldSize);
-
-      m_stream->flush();
-      m_stream->close();
-    }
-  }
-
-  /// Destructor flushes and closes the stream if still open.
-  ~EdfWriter() {
-    try {
-      close();
-    } catch (...) {
-      // Suppress exceptions in destructor.
-    }
-  }
-
-private:
-  explicit EdfWriter(std::unique_ptr<std::ofstream> stream,
-                     HeaderGeneral general)
-      : m_stream(std::move(stream)), m_general(std::move(general)) {}
-
-  std::unique_ptr<std::ofstream> m_stream;
-  HeaderGeneral m_general;
-  int64_t m_recordsWritten = 0;
 };
 
 } // namespace edfio
